Create embed_cluster_umap.py
This script scans the folder tiles_keep (this contains .tiff files taken from the whole slide image) which have been kept after quality control. This loads a pre-trained model called ResNet500 and resizes and normalises the tiles to be fed through the model. KMeans clustering is then performed to group the tile into N-clusters. UMAP embedding is done to reduce dimensionality to visualise the data.